+++ dorm-web/src/main/java/com/waken/dorm/controller/dorm/DormRuleController.java
@@ -15,6 +15,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
+import lombok.extern.slf4j.Slf4j;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,10 +27,10 @@ import org.springframework.web.bind.annotation.*;
  * @Author zhaoRong
  * @Date 2019/4/2 11:17
  **/
+@Slf4j
 @Api(value = "后台管理宿舍规则相关接口", description = "后台管理宿舍规则相关接口(AiShu)")
 @RestController
 public class DormRuleController extends BaseController {
-    private final Logger logger = LoggerFactory.getLogger(this.getClass());
     @Autowired
     private DormRuleService dormRuleService;
 
@@ -41,14 +42,8 @@ public class DormRuleController extends BaseController {
             @ApiResponse(code = 200, message = "success", response = DormRule.class)
     })
     public ResultView saveDormRule(@RequestBody EditDormRuleForm editDormRuleForm) {
-        logger.info("开始调用(保存/修改)宿舍规则信息接口接口：" + editDormRuleForm.toString());
-        try {
-            return ResultUtil.success(dormRuleService.saveDormRule(editDormRuleForm));
-        } catch (Exception e) {
-            e.printStackTrace();
-            logger.error("调用(保存/修改)宿舍规则信息接口失败:" + e.getMessage());
-            return ResultUtil.error();
-        }
+        log.info("开始调用(保存/修改)宿舍规则信息接口接口：" + editDormRuleForm.toString());
+        return ResultUtil.success(dormRuleService.saveDormRule(editDormRuleForm));
     }
 
     @Log("删除宿舍规则信息")
@@ -59,14 +54,12 @@ public class DormRuleController extends BaseController {
             @ApiResponse(code = 200, message = "success", response = ResultView.class)
     })
     public ResultView deleteDormRule(@RequestBody DeleteForm deleteFrom) {
-        logger.info("开始调用删除宿舍规则接口：" + deleteFrom.toString());
-        try {
-            dormRuleService.deleteDormRule(deleteFrom);
-            return ResultUtil.success();
-        } catch (Exception e) {
-            logger.error("调用删除宿舍规则接口失败:" + e.getMessage());
-            return ResultUtil.error();
+        log.info("开始调用删除宿舍规则接口：" + deleteFrom.toString());
+        if (null == deleteFrom.getDelIds() || deleteFrom.getDelIds().isEmpty()) {
+            return ResultUtil.errorByMsg("入参为空！");
         }
+        dormRuleService.deleteDormRule(deleteFrom);
+        return ResultUtil.success();
     }
 
     @CrossOrigin
@@ -76,13 +69,8 @@ public class DormRuleController extends BaseController {
             @ApiResponse(code = 200, message = "success", response = DormRuleView.class)
     })
     public ResultView listDormRules(@RequestBody DormRuleForm RuleForm) {
-        logger.info("开始调用分页查询宿舍规则信息接口：" + RuleForm.toString());
-        try {
-            RuleForm.setTerminal(CodeEnum.WEB.getCode());
-            return ResultUtil.success(dormRuleService.listDormRules(RuleForm));
-        } catch (Exception e) {
-            logger.info("调用分页查询宿舍规则信息失败:" + e.getMessage());
-            return ResultUtil.error();
-        }
+        log.info("开始调用分页查询宿舍规则信息接口：" + RuleForm.toString());
+        RuleForm.setTerminal(CodeEnum.WEB.getCode());
+        return ResultUtil.success(dormRuleService.listDormRules(RuleForm));
     }
 }
